fix misc compiler warnings
Siconos can now be compiled with others compilers (clang and icc)
This commit fixes a bunch of warnings, namely :
- implicit double -> conversion
- mixed class/struct declaration
- assert checking if an unisigned int is positive
- and more ...

Now we have a new macro to register a visitor for struct:
REGISTER_STRUCT

To compile Siconos with other compiler, you just need to set CC and CXX.
For instance:
export CC=clang CXX=clang++ for clang/llvm
export CC=icc CXX=icpc for ICC

For ICC, it is recommanded to use at least version 12.1. In this case,
you can pass -DICCOK=1 to cmake

git-svn-id: svn+ssh://gforge/svn/siconos/trunk@3269 bf01a9d7-0df3-0310-8ad4-b1c94d9ae2d5

+++ Numerics/src/NonSmoothSolvers/FrictionContact/FrictionContact3D_AlartCurnier.c
@@ -1446,7 +1446,7 @@ int DampedLocalNonsmoothNewtonSolver(FrictionContactProblem* localproblem, doubl
 
   double Tol = dparam[0];
   double itermax = iparam[0];
-  double LSitermax = iparam[1];
+  int LSitermax = iparam[1];
 
 
   int i, j, k, inew;
